+++ src/oprattr/__init__.py
@@ -1,20 +1,13 @@
 import collections.abc
 import functools
-import numbers
 
-from numerical import operators
 import numpy
 
 from . import abstract
+from . import methods
 from . import mixins
 from . import typeface
-from ._operations import (
-    unary,
-    equality,
-    ordering,
-    additive,
-    multiplicative,
-)
+from ._operations import equality
 
 
 T = typeface.TypeVar('T')
@@ -23,99 +16,31 @@ T = typeface.TypeVar('T')
 class Operand(abstract.Object[T], mixins.Numpy):
     """A concrete implementation of a real-valued object."""
 
-    def __abs__(self):
-        """Called for abs(self)."""
-        return unary(operators.abs, self)
+    __abs__ = methods.__abs__
+    __pos__ = methods.__pos__
+    __neg__ = methods.__neg__
 
-    def __pos__(self):
-        """Called for +self."""
-        return unary(operators.pos, self)
+    __eq__ = methods.__eq__
+    __ne__ = methods.__ne__
+    __lt__ = methods.__lt__
+    __le__ = methods.__le__
+    __gt__ = methods.__gt__
+    __ge__ = methods.__ge__
 
-    def __neg__(self):
-        """Called for -self."""
-        return unary(operators.neg, self)
-
-    def __eq__(self, other):
-        """Called for self == other."""
-        return equality(operators.eq, self, other)
-
-    def __ne__(self, other):
-        """Called for self != other."""
-        return equality(operators.ne, self, other)
-
-    def __lt__(self, other):
-        """Called for self < other."""
-        return ordering(operators.lt, self, other)
-
-    def __le__(self, other):
-        """Called for self <= other."""
-        return ordering(operators.le, self, other)
-
-    def __gt__(self, other):
-        """Called for self > other."""
-        return ordering(operators.gt, self, other)
-
-    def __ge__(self, other):
-        """Called for self >= other."""
-        return ordering(operators.ge, self, other)
-
-    def __add__(self, other):
-        """Called for self + other."""
-        return additive(operators.add, self, other)
-
-    def __radd__(self, other):
-        """Called for other + self."""
-        return additive(operators.add, other, self)
-
-    def __sub__(self, other):
-        """Called for self - other."""
-        return additive(operators.sub, self, other)
-
-    def __rsub__(self, other):
-        """Called for other - self."""
-        return additive(operators.sub, other, self)
-
-    def __mul__(self, other):
-        """Called for self * other."""
-        return multiplicative(operators.mul, self, other)
-
-    def __rmul__(self, other):
-        """Called for other * self."""
-        return multiplicative(operators.mul, other, self)
-
-    def __truediv__(self, other):
-        """Called for self / other."""
-        return multiplicative(operators.truediv, self, other)
-
-    def __rtruediv__(self, other):
-        """Called for other / self."""
-        return multiplicative(operators.truediv, other, self)
-
-    def __floordiv__(self, other):
-        """Called for self // other."""
-        return multiplicative(operators.floordiv, self, other)
-
-    def __rfloordiv__(self, other):
-        """Called for other // self."""
-        return multiplicative(operators.floordiv, other, self)
-
-    def __mod__(self, other):
-        """Called for self % other."""
-        return multiplicative(operators.mod, self, other)
-
-    def __rmod__(self, other):
-        """Called for other % self."""
-        return multiplicative(operators.mod, other, self)
-
-    def __pow__(self, other):
-        """Called for self ** other."""
-        if isinstance(other, numbers.Real):
-            return multiplicative(operators.pow, self, other)
-        return NotImplemented
-
-    def __rpow__(self, other):
-        """Called for other ** self."""
-        return NotImplemented
+    __add__ = methods.__add__
+    __radd__ = methods.__radd__
+    __sub__ = methods.__sub__
+    __rsub__ = methods.__rsub__
+    __mul__ = methods.__mul__
+    __rmul__ = methods.__rmul__
+    __truediv__ = methods.__truediv__
+    __rtruediv__ = methods.__rtruediv__
+    __floordiv__ = methods.__floordiv__
+    __rfloordiv__ = methods.__rfloordiv__
+    __mod__ = methods.__mod__
+    __rmod__ = methods.__rmod__
+    __pow__ = methods.__pow__
+    __rpow__ = methods.__rpow__
 
     def __array__(self, *args, **kwargs):
         """Called for numpy.array(self)."""